Remove and Sort using
1. Removed the Un-used using statements.
2. Sorted the Using statements.
3. Moved the using statements to the top in one file.
