+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/AddBasketItemRequest.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
+﻿namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
 {
+
     public class AddBasketItemRequest
     {
         public int CatalogItemId { get; set; }
+
         public string BasketId { get; set; }
 
         public int Quantity { get; set; }
@@ -17,4 +14,5 @@ namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
             Quantity = 1;
         }
     }
+
 }
