+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/BasketData.cs
@@ -2,6 +2,7 @@
 
 namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
 {
+
     public class BasketData
     {
         public string BuyerId { get; set; }
@@ -19,15 +20,4 @@ namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
         }
     }
 
-    public class BasketDataItem
-    {
-        public string Id { get; set; }
-        public int ProductId { get; set; }
-        public string ProductName { get; set; }
-        public decimal UnitPrice { get; set; }
-        public decimal OldUnitPrice { get; set; }
-        public int Quantity { get; set; }
-        public string PictureUrl { get; set; }
-
-    }
 }
