+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/UpdateBasketItemsRequest.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
 {
+
     public class UpdateBasketItemsRequest
     {
         
@@ -18,14 +16,4 @@ namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
         }
     }
 
-    public class UpdateBasketItemData
-    {
-        public string BasketItemId { get; set; }
-        public int NewQty { get; set; }
-
-        public UpdateBasketItemData()
-        {
-            NewQty = 0;
-        }
-    }
 }
