+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/UpdateBasketRequest.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
 {
+
     public class UpdateBasketRequest
     {
         public string BuyerId { get; set; }
@@ -12,10 +10,4 @@ namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
         public IEnumerable<UpdateBasketRequestItemData> Items { get; set; }
     }
 
-    public class UpdateBasketRequestItemData
-    {
-        public string Id { get; set; }          // Basket id
-        public int ProductId { get; set; }      // Catalog item id
-        public int Quantity { get; set; }       // Quantity
-    }
 }
