+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/UpdateBasketRequestItemData.cs
@@ -1,0 +1,13 @@
+﻿namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
+{
+
+    public class UpdateBasketRequestItemData
+    {
+        public string Id { get; set; }          // Basket id
+
+        public int ProductId { get; set; }      // Catalog item id
+
+        public int Quantity { get; set; }       // Quantity
+    }
+
+}
