+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Services/BasketService.cs
@@ -1,12 +1,11 @@
-﻿using Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Config;
+﻿using GrpcBasket;
+using Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Config;
 using Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using System.Threading.Tasks;
-using Grpc.Net.Client;
 using System.Linq;
-using GrpcBasket;
 using System.Net.Http;
+using System.Threading.Tasks;
 
 namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Services
 {
